requirements.txt
httpx>=0.27
selectolax>=0.3
google-genai>=1.0
resend>=2.0
python-dotenv>=1.0
